% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\docType{data}
\name{aaa}
\alias{aaa}
\title{DATASET_TITLE}
\format{
A data frame with 1938 rows and 87 variables:
\describe{
\item{\code{노선명}}{character COLUMN_DESCRIPTION}
\item{\code{도로.유형}}{character COLUMN_DESCRIPTION}
\item{\code{합계.길이}}{character COLUMN_DESCRIPTION}
\item{\code{소재지}}{character COLUMN_DESCRIPTION}
\item{\code{연번}}{integer COLUMN_DESCRIPTION}
\item{\code{자전거.전용도로}}{double COLUMN_DESCRIPTION}
\item{\code{자전거.보행자.겸용도로}}{double COLUMN_DESCRIPTION}
\item{\code{시군구}}{character COLUMN_DESCRIPTION}
\item{\code{시도}}{character COLUMN_DESCRIPTION}
\item{\code{자전거.차로.겸용도로}}{double COLUMN_DESCRIPTION}
\item{\code{연번}}{double COLUMN_DESCRIPTION}
\item{\code{도로유형}}{character COLUMN_DESCRIPTION}
\item{\code{광역자치단체}}{character COLUMN_DESCRIPTION}
\item{\code{시군구}}{character COLUMN_DESCRIPTION}
\item{\code{합계 길이}}{double COLUMN_DESCRIPTION}
\item{\code{관리부서}}{character COLUMN_DESCRIPTION}
\item{\code{기점}}{character COLUMN_DESCRIPTION}
\item{\code{노선명}}{character COLUMN_DESCRIPTION}
\item{\code{시점}}{character COLUMN_DESCRIPTION}
\item{\code{종점}}{character COLUMN_DESCRIPTION}
\item{\code{폭(m)}}{double COLUMN_DESCRIPTION}
\item{\code{포장종류}}{character COLUMN_DESCRIPTION}
\item{\code{도로종류}}{character COLUMN_DESCRIPTION}
\item{\code{자전거도로폭원(m)}}{character COLUMN_DESCRIPTION}
\item{\code{보도폭원(m)}}{character COLUMN_DESCRIPTION}
\item{\code{기준일자}}{double COLUMN_DESCRIPTION}
\item{\code{도로 유형}}{character COLUMN_DESCRIPTION}
\item{\code{도로연계성}}{character COLUMN_DESCRIPTION}
\item{\code{연장(km)}}{double COLUMN_DESCRIPTION}
\item{\code{도로부지}}{character COLUMN_DESCRIPTION}
\item{\code{하천부지}}{character COLUMN_DESCRIPTION}
\item{\code{자전거전용도로}}{double COLUMN_DESCRIPTION}
\item{\code{자전거차로겸용도로}}{double COLUMN_DESCRIPTION}
\item{\code{분리형겸용도로}}{double COLUMN_DESCRIPTION}
\item{\code{비분리형겸용도로}}{double COLUMN_DESCRIPTION}
\item{\code{보행량(점)}}{double COLUMN_DESCRIPTION}
\item{\code{도로명}}{character COLUMN_DESCRIPTION}
\item{\code{구간(기점)}}{character COLUMN_DESCRIPTION}
\item{\code{구간(종점)}}{character COLUMN_DESCRIPTION}
\item{\code{종류}}{character COLUMN_DESCRIPTION}
\item{\code{포장형태}}{character COLUMN_DESCRIPTION}
\item{\code{관리청}}{character COLUMN_DESCRIPTION}
\item{\code{자전거전용도로 비율(\%)}}{double COLUMN_DESCRIPTION}
\item{\code{자전거보행자겸용도로}}{double COLUMN_DESCRIPTION}
\item{\code{자전거보행자겸용도로 비율(\%)}}{double COLUMN_DESCRIPTION}
\item{\code{자전거전용차로}}{double COLUMN_DESCRIPTION}
\item{\code{자전거전용차로 비율(\%)}}{double COLUMN_DESCRIPTION}
\item{\code{자전거차로겸용도로  비율(%)}}{double COLUMN_DESCRIPTION}
  \item{\code{경유지}}{character COLUMN_DESCRIPTION}
  \item{\code{도로너비}}{character COLUMN_DESCRIPTION}
  \item{\code{자전거도로너비}}{character COLUMN_DESCRIPTION}
  \item{\code{교량개소수}}{double COLUMN_DESCRIPTION}
  \item{\code{교량전체길이(m)}}{double COLUMN_DESCRIPTION}
  \item{\code{안내표지수}}{double COLUMN_DESCRIPTION}
  \item{\code{노선번호}}{character COLUMN_DESCRIPTION}
  \item{\code{자전거전용도로연장}}{double COLUMN_DESCRIPTION}
  \item{\code{전용도로최소폭원(m)}}{double COLUMN_DESCRIPTION}
  \item{\code{전용도로최대폭원(m)}}{double COLUMN_DESCRIPTION}
  \item{\code{전용차로최소폭원(m)}}{double COLUMN_DESCRIPTION}
  \item{\code{전용차로최대폭원(m)}}{double COLUMN_DESCRIPTION}
  \item{\code{우선도로최소폭원(m)}}{double COLUMN_DESCRIPTION}
  \item{\code{우선도로최대폭원(m)}}{double COLUMN_DESCRIPTION}
  \item{\code{분리형겸용도로연장(km)}}{double COLUMN_DESCRIPTION}
  \item{\code{분리형겸용도로자전거도로최소폭원(m)}}{double COLUMN_DESCRIPTION}
  \item{\code{분리형겸용도로자전거도로최대폭원(m)}}{double COLUMN_DESCRIPTION}
  \item{\code{분리형겸용도로보도최소폭원(m)}}{double COLUMN_DESCRIPTION}
  \item{\code{분리형겸용도로보도최대폭원(m)}}{double COLUMN_DESCRIPTION}
  \item{\code{비분리형겸용도로연장(km)}}{double COLUMN_DESCRIPTION}
  \item{\code{...22}}{double COLUMN_DESCRIPTION}
  \item{\code{비분리형겸용도로최대폭원(m)}}{double COLUMN_DESCRIPTION}
  \item{\code{비분리형겸용도로유효최소폭원(m)}}{double COLUMN_DESCRIPTION}
  \item{\code{비분리형겸용도로유효최대폭원(m)}}{double COLUMN_DESCRIPTION}
  \item{\code{관리번호}}{character COLUMN_DESCRIPTION}
  \item{\code{도로관리청}}{character COLUMN_DESCRIPTION}
  \item{\code{도로폭원(m)}}{double COLUMN_DESCRIPTION}
  \item{\code{주요경유지}}{character COLUMN_DESCRIPTION}
  \item{\code{일반도로너비(m)}}{double COLUMN_DESCRIPTION}
  \item{\code{자전거도로너비(m)}}{character COLUMN_DESCRIPTION}
  \item{\code{사업량(km)}}{double COLUMN_DESCRIPTION}
  \item{\code{사업비(백만원)}}{double COLUMN_DESCRIPTION}
  \item{\code{비고}}{character COLUMN_DESCRIPTION}
  \item{\code{자전거 우선도로 길이}}{logical COLUMN_DESCRIPTION}
  \item{\code{너비(m)}}{double COLUMN_DESCRIPTION}
  \item{\code{지정사유(변경ㆍ폐지)}}{character COLUMN_DESCRIPTION}
  \item{\code{폭원(m)}}{double COLUMN_DESCRIPTION}
  \item{\code{자전거 보행자 겸용도로}}{double COLUMN_DESCRIPTION}
  \item{\code{자전거 전용도로}}{double COLUMN_DESCRIPTION}
}
}
\usage{
aaa
}
\description{
DATASET_DESCRIPTION
}
\details{
DETAILS
}
\keyword{datasets}
